[chore] dto 분리 및 파일 구조 정리
