+++ lib/src/components/page_graph.dart
@@ -1,6 +1,6 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:reafy_front/src/models/barData.dart';
+import 'package:reafy_front/src/dto/bar_dto.dart';
 
 class PageBarGraph extends StatefulWidget {
   final List monthlyPage;
