--- lib/src/models/bar.dart
@@ -1,6 +1,0 @@
-class IndividualBar {
-  final int x;
-  final double y;
-
-  IndividualBar({required this.x, required this.y});
-}
