+++ lib/src/provider/memo_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:reafy_front/src/repository/memo_repository.dart';
-import 'package:reafy_front/src/models/memo.dart';
+import 'package:reafy_front/src/dto/memo_dto.dart';
 
 class MemoProvider with ChangeNotifier {
   List<Memo> _memos = [];
