+++ lib/src/repository/item_repository.dart
@@ -3,25 +3,6 @@ import 'package:reafy_front/src/utils/api.dart';
 
 final Dio authdio = authDio().getDio();
 
-const url = 'https://dev.reafy.devkor.club';
-
-class ItemDto {
-  final int itemId;
-  bool activation;
-
-  ItemDto({
-    required this.itemId,
-    required this.activation,
-  });
-
-  factory ItemDto.fromJson(Map<String, dynamic> json) {
-    return ItemDto(
-      itemId: json['itemId'] as int,
-      activation: json['activation'] as bool,
-    );
-  }
-}
-
 //아이템 구매
 Future<bool> postItem(int itemId, bool activation, int price) async {
   ;
